moved Translations.js
